Fix flutter analyze linting issues
